+++ packages/pangea-sdk/pangea/asyncio/services/__init__.py
@@ -4,5 +4,5 @@ from .embargo import EmbargoAsync
 from .file_scan import FileScanAsync
 from .intel import DomainIntelAsync, FileIntelAsync, IpIntelAsync, UrlIntelAsync, UserIntelAsync
 from .redact import RedactAsync
-from .store import StoreAsync
+from .share import ShareAsync
 from .vault import VaultAsync