+++ packages/pangea-sdk/pangea/services/__init__.py
@@ -4,5 +4,5 @@ from .embargo import Embargo
 from .file_scan import FileScan
 from .intel import DomainIntel, FileIntel, IpIntel, UrlIntel, UserIntel
 from .redact import Redact
-from .store.store import Store
+from .share.share import Share
 from .vault.vault import Vault